refactor: sprint 40 — extract Package/Vulnerability/EPSS domains from _legacy
Phase 2 of the models split (audit item 1.1, rows 26-28). Following the
Sprint 39 pattern that extracted Image.

- libinv/models/package.py (new): Package, PackageLicenseAssociation,
  License.
- libinv/models/vulnerability.py (new): Vulnerability,
  VulnerabilityPackageAssociation.
- libinv/models/epss.py (new): EPSS.
- libinv/models/__init__.py: re-exports for all extracted classes so
  `from libinv.models import Package, Vulnerability, EPSS, …` keeps
  working unchanged.
- libinv/models/_legacy.py: shrinks; unused `Date`, `timedelta`
  imports dropped along with EPSS.

No test mock-patch redirections needed — extracted classes were only
patched as `libinv.models._legacy.X` for unrelated names
(DiscoveredPackage, session_scope, is_blacklist, is_excluded_repo,
Repository), none of which moved.

Verification: 90 passed, 3 skipped, 1 xfailed (baseline preserved
across all three increments). mypy libinv/models/ clean (7 source
files). `from libinv.models import Package, Vulnerability, EPSS,
PackageLicenseAssociation, License, VulnerabilityPackageAssociation`
smoke test prints ok.

Phase 3 (Sprint 41, rows 29-33) extracts the remaining classes:
Wasp/Actionable/Repository/Account/Secbug/Sast.
