+++ libinv/models/vulnerability.py
@@ -1,0 +1,107 @@
+"""Sprint 40.2: Vulnerability-domain ORM classes.
+
+Extracted from ``libinv/models/_legacy.py`` following the Sprint 39.2 /
+40.1 pattern. Holds the two classes describing the CVE-graph layer of
+the ORM:
+
+  * ``Vulnerability``                    — one upstream-feed CVE record
+                                            with severity + NVD CVSS
+                                            sub-scores. Description is
+                                            truncated to
+                                            ``MAX_LENGTH_VULNERABILITY_DESCRIPTION``
+                                            via ``set_description``.
+  * ``VulnerabilityPackageAssociation``  — M:N edge between Vulnerability
+                                            and Package, carrying a
+                                            comma-separated list of fix
+                                            versions in ``fix``.
+
+Per-relationship ``lazy=`` annotations were set under Sprint 37.2 and
+are preserved verbatim — see the inline comments on each
+``relationship(...)`` call for the audit trail.
+
+The package's previous home (`_legacy.py`) re-imports these names at the
+bottom of the file so any historical
+``from libinv.models._legacy import Vulnerability`` callers (and any
+internal helpers that referenced these names as module globals) keep
+working while the extraction is incremental.
+"""
+
+from __future__ import annotations
+
+from sqlalchemy import Column
+from sqlalchemy import Float
+from sqlalchemy import ForeignKey
+from sqlalchemy import Index
+from sqlalchemy import Integer
+from sqlalchemy import String
+from sqlalchemy.orm import relationship
+
+from libinv.base import Base
+
+# ``MAX_LENGTH_VULNERABILITY_DESCRIPTION`` remains canonically defined
+# in ``_legacy`` so the single-source-of-truth contract holds while the
+# rest of the file is still being split. Once the final extraction
+# sprint retires ``_legacy``, this constant will move to a shared
+# ``libinv.models._constants`` (TBD).
+from libinv.models._legacy import MAX_LENGTH_VULNERABILITY_DESCRIPTION
+
+
+class VulnerabilityPackageAssociation(Base):
+    __tablename__ = "vulnerability_package_association"
+    vulnerability_id = Column(
+        String(50),
+        ForeignKey("libinv.vulnerabilities.id", ondelete="CASCADE", onupdate="CASCADE"),
+        primary_key=True,
+    )
+    package_id = Column(
+        Integer,
+        ForeignKey("libinv.packages.id", ondelete="CASCADE", onupdate="CASCADE"),
+        primary_key=True,
+    )
+    fix = Column(String(100), doc="comma seperated list of fix versions", nullable=True)
+
+    # Sprint 37.2: lazy= audit.
+    # - vulnerability: traversed in image_scanner/sca.py via
+    #   selectinload(...VulnerabilityPackageAssociation.vulnerability). Keep select.
+    # - package: back-ref never traversed in api/cli/scanners.
+    vulnerability = relationship("Vulnerability", back_populates="packages", lazy="select")
+    package = relationship("Package", back_populates="vulnerabilities", lazy="raise_on_sql")
+
+    # Sprint 33.1/33.2: declare indexes already created by alembic 0002_fk_indexes
+    __table_args__ = (
+        Index("ix_vulnerability_package_association_package_id", "package_id"),
+        {"schema": "libinv"},
+    )
+
+
+class Vulnerability(Base):
+    __tablename__ = "vulnerabilities"
+
+    id = Column(String(50), primary_key=True)
+    # Sprint 34.1: upstream-feed-derived fields — feeds may omit any of them.
+    description = Column(String(MAX_LENGTH_VULNERABILITY_DESCRIPTION), nullable=True)
+    severity = Column(String(10), nullable=True)
+    related = Column(
+        String(200), doc="comma seperated list of related cve ids", nullable=True
+    )
+    nvd_cvss_base_score = Column(
+        "nvd-cvss.base_score", Float(precision=3), nullable=True
+    )
+    nvd_cvss_exploitability_score = Column(
+        "nvd-cvss.exploitability_score", Float(precision=3), nullable=True
+    )
+    nvd_cvss_impact_score = Column(
+        "nvd-cvss.impact_score", Float(precision=3), nullable=True
+    )
+    # Sprint 37.2: back-ref never traversed in api/cli/scanners (only the forward
+    # VulnerabilityPackageAssociation.vulnerability direction is read).
+    packages = relationship(
+        "VulnerabilityPackageAssociation", back_populates="vulnerability", lazy="raise_on_sql"
+    )
+
+    def set_description(self, desc: str):
+        if desc:
+            self.description = desc[:MAX_LENGTH_VULNERABILITY_DESCRIPTION]
+
+    def __str__(self):
+        return self.id
